projects/ambient-email-agent/pyproject.toml
[project]
name = "ambient-email-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "langgraph-cli[inmem]>=0.4.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "langsmith>=0.5.1",
    "pytest>=9.0.2",
]
